Stray newlines in comments and datatype change for iterator variable in getColumnIndex

+++ tightdb-java-core/src/main/java/com/tightdb/TableView.java
@@ -111,8 +111,7 @@ public class TableView implements TableOrView {
      * Returns the name of a column identified by columnIndex. Notice that the
      * index is zero based.
      *
-     * @param columnIndex
-     *            the column index
+     * @param columnIndex the column index
      * @return the name of the column
      */
     public String getColumnName(long columnIndex) {
@@ -124,13 +123,12 @@ public class TableView implements TableOrView {
     /**
      * Returns the 0-based index of a column based on the name.
      *
-     * @param name
-     *            column name
+     * @param name column name
      * @return the index, -1 if not found
      */
     public long getColumnIndex(String name) {
         long columnCount = getColumnCount();
-        for (int i = 0; i < columnCount; i++) {
+        for (long i = 0; i < columnCount; i++) {
             if (name.equals(getColumnName(i))) {
                 return i;
             }
@@ -141,8 +139,7 @@ public class TableView implements TableOrView {
     /**
      * Get the type of a column identified by the columnIdex.
      *
-     * @param columnIndex
-     *            index of the column.
+     * @param columnIndex index of the column.
      * @return Type of the particular column.
      */
     public ColumnType getColumnType(long columnIndex)
